% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplefinr-package.R
\docType{package}
\name{simplefinr-package}
\alias{simplefinr}
\alias{simplefinr-package}
\title{simplefinr: Query SimpleFIN Server APIs and Parse Financial Data}
\description{
Provides tools for claiming SimpleFIN tokens, querying a SimpleFIN server or bridge API, and parsing the JSON response into flattened tibbles for accounts, transactions, connections, and errors. Implements the SimpleFIN Protocol v2.0.
}
\author{
\strong{Maintainer}: Dan Riggi

Authors:
\itemize{
  \item Dan Riggi
}

}
\keyword{internal}
